+++ oracledbStorage/src/test/java/io/github/factoryfx/factory/datastorage/oracle/OracledbDataStorageHistoryTest.java
@@ -16,14 +16,14 @@ public class OracledbDataStorageHistoryTest extends DatabaseTest {
 
     @Test
     public void test_empty() {
-        OracledbDataStorageHistory<ExampleFactoryA> oracledbDataStorageHistory = new OracledbDataStorageHistory<>(connectionSupplier, createMigrationManager());
+        OracledbDataStorageHistory<ExampleFactoryA> oracledbDataStorageHistory = new OracledbDataStorageHistory<>(connectionSupplier, createMigrationManager(), false);
 
         Assertions.assertTrue(oracledbDataStorageHistory.getHistoryFactoryList().isEmpty());
     }
 
     @Test
     public void test_add() {
-        OracledbDataStorageHistory<ExampleFactoryA> oracledbDataStorageHistory = new OracledbDataStorageHistory<>(connectionSupplier, createMigrationManager());
+        OracledbDataStorageHistory<ExampleFactoryA> oracledbDataStorageHistory = new OracledbDataStorageHistory<>(connectionSupplier, createMigrationManager(), false);
 
         StoredDataMetadata metadata = createDummyMetadata();
         oracledbDataStorageHistory.updateHistory(metadata,new ExampleFactoryA());
@@ -41,7 +41,7 @@ public class OracledbDataStorageHistoryTest extends DatabaseTest {
 
     @Test
     public void test_multi_add() {
-        OracledbDataStorageHistory<ExampleFactoryA> oracledbDataStorageHistory = new OracledbDataStorageHistory<>(connectionSupplier, createMigrationManager());
+        OracledbDataStorageHistory<ExampleFactoryA> oracledbDataStorageHistory = new OracledbDataStorageHistory<>(connectionSupplier, createMigrationManager(), false);
 
         {
             oracledbDataStorageHistory.updateHistory(createDummyMetadata(), new ExampleFactoryA());
@@ -60,23 +60,23 @@ public class OracledbDataStorageHistoryTest extends DatabaseTest {
 
     @Test
     public void test_restore() {
-        OracledbDataStorageHistory<ExampleFactoryA> oracledbDataStorageHistory = new OracledbDataStorageHistory<>(connectionSupplier, createMigrationManager());
+        OracledbDataStorageHistory<ExampleFactoryA> oracledbDataStorageHistory = new OracledbDataStorageHistory<>(connectionSupplier, createMigrationManager(), false);
 
         oracledbDataStorageHistory.updateHistory(createDummyMetadata(),new ExampleFactoryA());
         assertEquals(1, oracledbDataStorageHistory.getHistoryFactoryList().size());
 
-        OracledbDataStorageHistory<ExampleFactoryA> restored = new OracledbDataStorageHistory<>(connectionSupplier, createMigrationManager());
+        OracledbDataStorageHistory<ExampleFactoryA> restored = new OracledbDataStorageHistory<>(connectionSupplier, createMigrationManager(), false);
         assertEquals(1,restored.getHistoryFactoryList().size());
     }
 
     @Test
     public void test_getById() {
-        OracledbDataStorageHistory<ExampleFactoryA> oracledbDataStorageHistory = new OracledbDataStorageHistory<>(connectionSupplier, createMigrationManager());
+        OracledbDataStorageHistory<ExampleFactoryA> oracledbDataStorageHistory = new OracledbDataStorageHistory<>(connectionSupplier, createMigrationManager(), false);
 
         oracledbDataStorageHistory.updateHistory(createDummyMetadata(),new ExampleFactoryA());
         assertEquals(1, oracledbDataStorageHistory.getHistoryFactoryList().size());
 
-        OracledbDataStorageHistory<ExampleFactoryA> history = new OracledbDataStorageHistory<>(connectionSupplier, createMigrationManager());
+        OracledbDataStorageHistory<ExampleFactoryA> history = new OracledbDataStorageHistory<>(connectionSupplier, createMigrationManager(), false);
 
         ExampleFactoryA reloaded = oracledbDataStorageHistory.getHistoryFactory(new ArrayList<>(history.getHistoryFactoryList()).get(0).id);
         Assertions.assertNotNull(reloaded);